cleanup, moving packages, __job_stub updated
