+++ _agent/events/Events.py
@@ -1,5 +1,4 @@
 import logging
-from abc import ABC, abstractmethod
 
 
 class Publisher:
