+++ _agent/jobs/FindService.py
@@ -4,12 +4,29 @@ from dbus import Interface
 from _agent.events.Events import Publisher
 from _agent.events.EventsType import EventsType
 from _agent.manager import Sysd
-from _agent.scheduler.Scheduler import Job
+from _agent.jobs.scheduler import Job
 
 
 class UnitNotFoundException(Exception):
     def __init__(self, obj):
         self.obj = obj
+
+
+def _find_service_unit(loop: bool, callback: callable, fallback: callable, service_name: str):
+    """Find the service unit by it's name.
+        :param:
+            `name`:`the formatted service name as {name}.service`
+            `loop`:`true or false to make it loop`
+        :returns:
+            `service_object_path`:`the service object path reference`
+    """
+    Sysd.get_manager().GetUnit(
+        service_name,
+        reply_handler=callback,
+        error_handler=fallback
+    )
+    # return false to not loop
+    return loop
 
 
 class FindUnitPropertiesJob(Job):
@@ -39,22 +56,3 @@ class FindUnitPropertiesJob(Job):
         print(f"async client {self} status: ExceptionRaise {e}")
         # if an error happens on read i don't need to quit
         # loop.quit()
-
-
-def _find_service_unit(loop: bool, callback: callable, fallback: callable, service_name: str):
-    """Find the service unit by it's name.
-        :param:
-            `name`:`the formatted service name as {name}.service`
-            `loop`:`true or false to make it loop`
-        :returns:
-            `service_object_path`:`the service object path reference`
-    """
-    Sysd.get_manager().GetUnit(
-        service_name,
-        reply_handler=callback,
-        error_handler=fallback
-    )
-    # return false to not loop
-    return loop
-
-
