+++ _agent/models/RestartServiceParameters.py
@@ -1,0 +1,7 @@
+class RestartServiceParams:
+
+    def __init__(self,
+                 service_name: str,
+                 mode: str = 'replace'):
+        self.service_name = service_name
+        self.mode = mode
